♻️ Refactor: 디데이 생성에 필요한 상태 커스텀 훅으로 변환
refactor: 디데이 생성에 필요한 상태 커스텀 훅으로 변환
