+++ src/pages/dday/NewDday.tsx
@@ -1,69 +1,29 @@
 // 외부 라이브러리
-import { useState, useMemo } from "react";
 import { useNavigate } from "react-router";
 
 // UI 컴포넌트
 import { Button, Divider, Topbar } from "@/shared/ui";
-import { DdayDateBottomSheet, EmojiSelector } from "@/features/d-day/ui";
+import { DdayDateBottomSheet, EmojiSelector } from "@/features/d-day";
 import { TitleInput } from "@/features/schedule";
 
 // 아이콘
 import backIcon from "@/assets/icons/back.svg";
 
-// 타입 및 API
+// API 및 훅
 import { useDdayMutation } from "@/entities/d-day";
-import { Emoji } from "@/entities/d-day";
-// 타입 정의
-interface InitialDday {
-  id: number | null;
-  title: string;
-  date: string;
-  emoji: Emoji;
-}
+import { useNewDdayForm } from "@/features/d-day";
 
 export const NewDday = () => {
   // 라우터 훅
   const navigate = useNavigate();
 
-  // 상태
-  const initialState: InitialDday = {
-    id: null,
-    title: "",
-    date: "",
-    emoji: Emoji.HEART,
-  };
-  const [newDdayState, setNewDdayState] = useState<InitialDday>(initialState);
+  // 상태 및 폼 관련 커스텀 훅
+  const { newDdayState, handleChange, isFormValid } = useNewDdayForm();
 
   // API 훅
   const { mutate: ddayMutate } = useDdayMutation(newDdayState);
 
-  // 유효성 검사
-  const isFormValid = useMemo(() => {
-    return newDdayState.title.trim() !== "" && newDdayState.date !== "";
-  }, [newDdayState.title, newDdayState.date]);
-
   // 이벤트 핸들러
-  const handleTitleChange = (title: string) => {
-    setNewDdayState(prev => ({
-      ...prev,
-      title,
-    }));
-  };
-
-  const handleDateChange = (date: string) => {
-    setNewDdayState(prev => ({
-      ...prev,
-      date,
-    }));
-  };
-
-  const handleEmojiChange = (emoji: Emoji) => {
-    setNewDdayState(prev => ({
-      ...prev,
-      emoji,
-    }));
-  };
-
   const handlePostDday = async () => {
     ddayMutate(undefined, {
       onSuccess: () => {
@@ -81,7 +41,6 @@ export const NewDday = () => {
     return goBack;
   };
 
-  // 렌더링
   return (
     <>
       <header className="mt-[70px] mb-8 flex flex-col items-center gap-7">
@@ -95,17 +54,17 @@ export const NewDday = () => {
       </header>
       <main className="flex flex-col gap-6 p-5">
         <section className="flex flex-col gap-2">
-          <TitleInput value={newDdayState.title} onTitleChange={handleTitleChange} />
+          <TitleInput value={newDdayState.title} onTitleChange={value => handleChange("title", value)} />
         </section>
         <Divider thickness="h-px" color="bg-gray-100" />
         <section className="flex flex-col gap-4">
           <div className="flex flex-col gap-2">
             <h3 className="text-xl font-semibold text-gray-900">이모티콘</h3>
-            <EmojiSelector selectedEmoji={newDdayState.emoji} onEmojiChange={handleEmojiChange} />
+            <EmojiSelector selectedEmoji={newDdayState.emoji} onEmojiChange={value => handleChange("emoji", value)} />
           </div>
           <div className="flex flex-col gap-2">
             <h3 className="text-xl font-semibold text-gray-900">날짜</h3>
-            <DdayDateBottomSheet selectedDate={newDdayState.date} onDateChange={handleDateChange} />
+            <DdayDateBottomSheet selectedDate={newDdayState.date} onDateChange={value => handleChange("date", value)} />
           </div>
         </section>
         <section className="fixed bottom-6 left-1/2 w-[375px] -translate-x-1/2 transform px-4">
